vdW interactions and thermal radiation functions

Thermal radiation

heattransfer — Function

heattransfer(myARGS)

Extract command-line arguments for computing the radiative heat transfer among a collection of molecules, then compute the radiative heat transfer appropriately.

Note

Any frequencies will be converted to their absolute values. Frequencies are not allowed to be exactly zero.

Note

This computes the integrand $\Phi^{(m)}_{n}$ at specified real frequencies, without doing any integration. This includes neither numerical prefactors outside of the trace, nor the Planck function.

Tip

This function can be embarrassingly parallelized over frequency and wavevector. However, the integrand at each frequency and wavevector is calculated for every transformation, as this maximizes the efficiency of the program by virtue of reusing quantities independent of geometric transformations.

heattransfer_general — Function

heattransfer_general(outfilename, myAllMolData, myTransData, myPeriodicData,
                     myDeltaKISystemData, Genvstr, nonretarded, freqklist)

Compute the actual heat transfer integrands (for conduction, radiation, and both together) and print each to appropriate files based on the name specified by outfilename.

printPhimn — Function

printPhimn(Phimn::Array{FT, 2}, outfilename::AbstractString,
           myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
           myTransData::TransData{FT},
           myPeriodicData::PeriodicData{FT},
           tt::Integer, currfreq::Union{FT, Complex{FT}},
           currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Print the matrix Phimn to the file with name outfilename. The elements Phimn[m, n] represent the energy transfer from molecule m to molecule n, corresponding to the mathematical quantity $\Phi^{(m)}_{n}$. If there is no spatial periodicity, Phimn should be symmetric, and a line of output will be as follows: myTransData.transLabels[tt] currfreq Phimn[1, 1] Phimn[1, 2] ... Phimn[2, 1] Phimn[2, 2] ... for example: TRANS05 1.2340953873461249e+12 -6.3540963812489034e-01 6.3540963812489034e-01 6.3540963812489034e-01 -6.3540963812489034e-01 If there is spatial periodicity, the transpose of Phimn at wavevector $ec{k}$ is the equivalent of computing Phimn at $- ec{k}$, and a line of output will be as follows: myTransData.transLabels[tt] currfreq currk[1] currk[2] currk[3] Phimn[1, 1] Phimn[1, 2] ... Phimn[2, 1] Phimn[2, 2] ... for example: TRANS05 1.2340953873461249e+12 5.0000000000000000e+07 3.0000000000000000e+07 0.0000000000000000e+00 -6.3540963812489034e-01 6.3540963812489034e-01 6.3540963812489034e-01 -6.3540963812489034e-01

assembleZIIinvnoDeltaKI! — Function

assembleZIIinvnoDeltaKI!(totalZIIinvmat::SharedArray{Complex{FT}, 2},
                         myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                         myTransData::TransData{FT},
                         tt::Integer)

Assemble the matrix totalZIIinvmat in-place for the geometrical transformation with index tt, in the absence of internuclear couplings between different molecules.

assembleZIIinvTeeradiation! — Function

assembleZIIinvTeeradiation!(totalZIIinvmat::SharedArray{Complex{FT}, 2},
                            totalTeemat::SharedArray{Complex{FT}, 2},
                            totalGenvmat::SharedArray{Complex{FT}, 2},
                            myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                            myTransData::TransData{FT},
                            tt::Integer)

Assemble the matrices totalZIIinvmat and totalTeemat in-place for the geometrical transformation with index tt, in the absence of internuclear couplings between different molecules but in the presence of radiative couplings.

assembletempmateI! — Function

assembletempmateI!(tempmat::SharedArray{Complex{FT}, 2},
                   totalTeemat::SharedArray{Complex{FT}, 2},
                   totalZIIinvmat::SharedArray{Complex{FT}, 2},
                   myAllMolData::MolSystem{OneMol_WithPhonons{FT}})

Assemble the matrix tempmat in-place, equal to $(Z_{\mathrm{ee}} - K_{\mathrm{e}} Z_{\mathrm{II}}^{-1} K_{\mathrm{e}})^{-1} K_{\mathrm{e}} Z_{\mathrm{II}}^{-1}$.

assemblePhimnwithBe! — Function

assemblePhimnwithBe!(Phimn::Array{FT, 2},
                     myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                     myPeriodicData::PeriodicData{FT},
                     tempmat::SharedArray{Complex{FT}, 2},
                     DeltaZImat::SharedArray{Complex{FT}, 2},
                     currfreq::Union{FT, Complex{FT}})

Assemble the contributions to the matrix Phimn in-place, corresponding to the fluctuating electronic sources (accounting for absorption in both electrons due to radiation and nuclei due to conduction); contributions are added to the existing values of Phimn.

assemblePhimnwithBI! — Function

assemblePhimnwithBI!(Phimn::Array{FT, 2},
                     myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                     myPeriodicData::PeriodicData{FT},
                     tempmat::SharedArray{Complex{FT}, 2},
                     DeltaZImat::SharedArray{Complex{FT}, 2},
                     currfreq::Union{FT, Complex{FT}})

Assemble the contributions to the matrix Phimn in-place, corresponding to the fluctuating nuclear sources (accounting for absorption in both electrons due to radiation and nuclei due to conduction); contributions are added to the existing values of Phimn.

assemblePhimnradiation! — Function

assemblePhimnradiation!(Phimn::Array{FT, 2},
                        myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                        myPeriodicData::PeriodicData{FT},
                        totalTeemat::SharedArray{Complex{FT}, 2},
                        totalGenvmatwithdiag::SharedArray{Complex{FT}, 2},
                        totalZIIinvmat::SharedArray{Complex{FT}, 2},
                        tempmat::SharedArray{Complex{FT}, 2},
                        currfreq::Union{FT, Complex{FT}})

Assemble the matrix Phimn in-place for purely radiative couplings.

assembleZIIinvwithDeltaKI! — Function

assembleZIIinvwithDeltaKI!(totalZIIinvmat::SharedArray{Complex{FT}, 2},
                           DeltaKImat::SharedArray{Complex{FT}, 2},
                           myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                           myTransData::TransData{FT},
                           tt::Integer)

Assemble the matrix totalZIIinvmat in-place for the geometrical transformation with index tt, in the presence of internuclear couplings between different molecules.

assembleZIIinvTeeconduction! — Function

assembleZIIinvTeeconduction!(totalZIIinvmat::SharedArray{Complex{FT}, 2},
                             totalTeemat::SharedArray{Complex{FT}, 2},
                             DeltaKImat::SharedArray{Complex{FT}, 2},
                             myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                             myTransData::TransData{FT},
                             tt::Integer)

Assemble the matrices totalZIIinvmat and totalTeemat in-place for the geometrical transformation with index tt, in the presence of internuclear couplings between different molecules but in the absence of radiative couplings.

assembletempmatIe! — Function

assembletempmatIe!(tempmat::SharedArray{Complex{FT}, 2},
                   totalTeemat::SharedArray{Complex{FT}, 2},
                   totalZIIinvmat::SharedArray{Complex{FT}, 2},
                   myAllMolData::MolSystem{OneMol_WithPhonons{FT}})

Assemble the matrix tempmat in-place, equal to $Z_{\mathrm{II}}^{-1} K_{\mathrm{e}} (Z_{\mathrm{ee}} - K_{\mathrm{e}} Z_{\mathrm{II}}^{-1} K_{\mathrm{e}})^{-1}$.

assembletempmatII! — Function

assembletempmatII!(tempmat::SharedArray{Complex{FT}, 2},
                   totalTeemat::SharedArray{Complex{FT}, 2},
                   totalZIIinvmat::SharedArray{Complex{FT}, 2},
                   myAllMolData::MolSystem{OneMol_WithPhonons{FT}})

Assemble the matrix tempmat in-place, equal to $Z_{\mathrm{II}}^{-1} + Z_{\mathrm{II}}^{-1} K_{\mathrm{e}} (Z_{\mathrm{ee}} - K_{\mathrm{e}} Z_{\mathrm{II}}^{-1} K_{\mathrm{e}})^{-1} K_{\mathrm{e}} Z_{\mathrm{II}}^{-1}$.

assemblePhimnconduction! — Function

assemblePhimnconduction!(Phimn::Array{FT, 2},
                         myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                         myPeriodicData::PeriodicData{FT},
                         totalTeemat::SharedArray{Complex{FT}, 2},
                         DeltaKImat::SharedArray{Complex{FT}, 2},
                         totalZIIinvmat::SharedArray{Complex{FT}, 2},
                         tempmat::SharedArray{Complex{FT}, 2},
                         currfreq::Union{FT, Complex{FT}})

Assemble the matrix Phimn in-place for purely conductive couplings.

assembleZIIinvTeeboth! — Function

assembleZIIinvTeeboth!(totalZIIinvmat::SharedArray{Complex{FT}, 2},
                       totalTeemat::SharedArray{Complex{FT}, 2},
                       totalGenvmat::SharedArray{Complex{FT}, 2},
                       DeltaKImat::SharedArray{Complex{FT}, 2},
                       myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                       myTransData::TransData{FT},
                       tt::Integer)

Assemble the matrices totalZIIinvmat and totalTeemat in-place for the geometrical transformation with index tt, in the presence of both radiative and internuclear couplings between different molecules.

assemblePhimnboth! — Function

assemblePhimnboth!(Phimn::Array{FT, 2},
                   myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                   myPeriodicData::PeriodicData{FT},
                   totalTeemat::SharedArray{Complex{FT}, 2},
                   totalGenvmatwithdiag::SharedArray{Complex{FT}, 2},
                   DeltaKImat::SharedArray{Complex{FT}, 2},
                   totalZIIinvmat::SharedArray{Complex{FT}, 2},
                   tempmat::SharedArray{Complex{FT}, 2},
                   currfreq::Union{FT, Complex{FT}})

Assemble the matrix Phimn in-place for both radiative and conductive couplings.

assembleGenvDiagBlock! — Function

assembleGenvDiagBlock!(totalGenvmat::SharedArray{Complex{FT}, 2},
                       totalGenvmatwithdiag::SharedArray{Complex{FT}, 2},
                       myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                       myTransData::TransData{FT}, nn2::Integer,
                       currmolind2::Integer, startidx2::Integer,
                       tt::Integer, myPeriodicData::PeriodicData{FT},
                       GFSCAGG!::Function,
                       currfreqG::Union{FT, Complex{FT}},
                       currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular diagonal block of the scattering Green's function interaction matrix of molecule nn2 (whose unique molecular index currmolind2 is precomputed and passed as an argument) at frequency currfreqG and wavevector currk using GFSCAGG!, then add the vacuum contribution of molecule nn2, performing any required matrix rotations corresponding to transformation tt for molecule nn2. Stamp that into totalGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])], and added to coincident contributions into totalGenvmatwithdiag[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Finally, undo any rotation corresponding to transformation tt in molecule nn2 for reuse with future transformations. This assumes the presence of phonons.

assembleGenvOffDiagBlock! — Function

assembleGenvOffDiagBlock!(totalGenvmat::SharedArray{Complex{FT}, 2},
                          totalGenvmatwithdiag::SharedArray{Complex{FT}, 2},
                          myAllMolData::MolSystem{<:OneMol{FT}},
                          myTransData::TransData{FT},
                          nn1::Integer, nn2::Integer,
                          startidx1::Integer,
                          startidx2::Integer,
                          myPeriodicData::PeriodicData{FT},
                          GFSCAGG!::Function,
                          currfreqG::Union{FT, Complex{FT}},
                          currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular off-diagonal block of the scattering Green's function interaction matrix of molecules nn2 as the source and nn1 as the field (whose respective unique molecular indices currmolind1 and currmolind2 are precomputed and passed as arguments) at frequency currfreqG and wavevector currk using GFSCAGG!, multiply on the left by the electronic polarizabilities, add the vacuum contribution, and stamp that into totalGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])] and totalGenvmatwithdiag[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. If there is no spatial periodicity, do the same for the (unconjugated) transpose, swapping nn2 and nn1 without performing further Green's function computations.

vdW interactions

vdWenergy — Function

vdWenergy(myARGS)

Extract command-line arguments for computing the vdW interaction energy among a collection of molecules, then compute the vdW energy integrand in imaginary frequency appropriately.

Note

Any frequencies will be converted to their absolute values times 1im. Frequencies are not allowed to be exactly zero.

Note

This computes the integrand $\ln(\det(T_{\infty} T^{-1}))$ at specified imaginary frequencies, without doing any integration or Matsubara summation. This is typically a negative dimensionless quantity, and no factors of $2\pi$ or $\hbar$ are included.

Tip

This function can be embarrassingly parallelized over frequency and wavevector. However, the integrand at each frequency and wavevector is calculated for every transformation, as this maximizes the efficiency of the program by virtue of reusing quantities independent of geometric transformations.

Tip

The recommended way to do frequency integration or Matsubara summation is to supply a list of frequencies (typically not too many), and then interpolate the integrand to do integration or Matsubara summation after the fact.

vdWenergy_general — Function

vdWenergy_general(outfilename, myAllMolData, myTransData, myPeriodicData,
                  myDeltaKISystemData, Genvstr, nonretarded, freqklist)

Compute the actual vdW energy integrands (for conduction, radiation, and both together) and print each to outfilename (radiation integrand, then conduction integrand, then total integrand).

constructMolZI0! — Function

constructMolZI0!(myMolData::OneMol_WithPhonons{FT},
                 myPeriodicData::PeriodicData{FT},
                 freq::Union{FT, Complex{FT}},
                 k::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Construct the product of the uncoupled nuclear response with the internuclear spring constant matrix of a single molecule. This also constructs the atomic polarizabilities used to evaluate the Gaussian widths used for the Green's function matrices.

constructMolYe0eeGvacinf! — Function

constructMolYe0eeGvacinf!(myMolData::OneMol_WithPhonons{FT},
                          myPeriodicData::PeriodicData{FT},
                          currfreqG::Union{FT, Complex{FT}},
                          currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the Green's function interaction matrix of a molecule in vacuum myMolData.GvacGGinf, then multiplies on the left by its electronic response in-place, at frequency currfreqG and wavevector currk passing myMolData and myPeriodicData to the subsidiary functions. This matrix is then subtracted from the identity.

computeradintegrandvac — Function

computeradintegrandvac(tempmat::SharedArray{Complex{FT}, 2},
                       myMolData::OneMol_WithPhonons{FT}) where FT<:AbstractFloat

Set up the matrix tempmat to compute the vacuum radiative contribution to the vdW energy integrand.

integrandvac — Function

integrandvac(myAllMolData::MolSystem{<:OneMol{FT}},
             integrandvaclist::Array{FT, 1}) where FT<:AbstractFloat

Populate the list of vacuum contributions to the vdW integrand integrandvaclist with contributions from duplicate molecules where they exist, then return the sum yielding the total contribution to the integrand in vacuum.

assembleYe0eeGenvDiagBlock! — Function

assembleYe0eeGenvDiagBlock!(totalYe0eeGenvmat::SharedArray{Complex{FT}, 2},
                            myAllMolData::MolSystem{<:OneMol{FT}},
                            myTransData::TransData{FT}, nn2::Integer,
                            currmolind2::Integer, startidx2::Integer,
                            tt::Integer, myPeriodicData::PeriodicData{FT},
                            GFSCAGG!::Function,
                            currfreqG::Union{FT, Complex{FT}},
                            currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular diagonal block of the scattering Green's function interaction matrix of molecule nn2 (whose unique molecular index currmolind2 is precomputed and passed as an argument) at frequency currfreqG and wavevector currk using GFSCAGG!, and multiply on the left by the electronic response, then add the vacuum contribution which has already been computed as the electronic polarizabilities multiplied by the vacuum Green's function matrix of molecule nn2, performing any required matrix rotations corresponding to transformation tt for molecule nn2, and stamp that into totalYe0eeGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Finally, undo any rotation corresponding to transformation tt in the vacuum Green's function matrix of molecule nn2 for reuse with future transformations.

assembleYe0eeGenvOffDiagBlock! — Function

assembleYe0eeGenvOffDiagBlock!(totalYe0eeGenvmat::SharedArray{Complex{FT}, 2},
                               myAllMolData::MolSystem{<:OneMol{FT}},
                               myTransData::TransData{FT},
                               nn1::Integer, nn2::Integer,
                               startidx1::Integer,
                               startidx2::Integer,
                               myPeriodicData::PeriodicData{FT},
                               GFSCAGG!::Function,
                               currfreqG::Union{FT, Complex{FT}},
                               currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular off-diagonal block of the scattering Green's function interaction matrix of molecules nn2 as the source and nn1 as the field (whose respective unique molecular indices currmolind1 and currmolind2 are precomputed and passed as arguments) at frequency currfreqG and wavevector currk using GFSCAGG!, multiply on the left by the electronic response, add the vacuum contribution, and stamp that into totalYe0eeGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Do the same for the adjoint, swapping nn2 and nn1 without performing further Green's function computations.

assembleYe0IIDeltaKI! — Function

assembleYe0IIDeltaKI!(Ye0IIDeltaKImat::AbstractArray{Complex{FT}, 2},
                      myDeltaKISystemData::DeltaKISystemData{FT},
                      myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                      blochk::AbstractArray{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Construct the product of the uncoupled nuclear response across all molecules with the internuclear spring constant matrix coupling different molecules.

computeradintegrandnoDeltaKI — Function

computeradintegrandnoDeltaKI(tempmat::SharedArray{Complex{FT}, 2},
                             totalYe0eeGenvmat::SharedArray{Complex{FT}, 2},
                             myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                             myTransData::TransData{FT},
                             tt::Integer) where FT<:AbstractFloat

Compute and return the total contribution (before subtracting the vacuum contribution) to the vdW energy integrand from radiation in the absence of conductive couplings between different molecules, at transformation index tt.

computeradintegrandwithDeltaKI — Function

computeradintegrandwithDeltaKI(tempmat::SharedArray{Complex{FT}, 2},
                               totalYe0eeGenvmat::SharedArray{Complex{FT}, 2},
                               Ye0IIDeltaKImat::SharedArray{Complex{FT}, 2},
                               myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                               myTransData::TransData{FT},
                               tt::Integer) where FT<:AbstractFloat

Compute and return the total contribution (before subtracting the vacuum contribution) to the vdW energy integrand from radiation in the presence of conductive couplings between different molecules, at transformation index tt.

computecondintegrandwithDeltaKI — Function

computecondintegrandwithDeltaKI(tempmat::SharedArray{Complex{FT}, 2},
                                Ye0IIDeltaKImat::SharedArray{Complex{FT}, 2},
                                myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                                myTransData::TransData{FT},
                                tt::Integer) where FT<:AbstractFloat

Compute and return the total contribution (with the vacuum contribution vanishing) to the vdW energy integrand from conduction (in the presence of conductive couplings between different molecules), at transformation index tt.