Intermolecular internuclear coupling data structures and methods

Data structures

DeltaKIMolBlockData — Type

DeltaKIMolBlockData{FT<:AbstractFloat}

Represent the internuclear couplings between a pair of molecules, assumed to be computed via a Morse potential, with the possibility of Bloch periodicity.

See also: DeltaKISystemData{FT}

DeltaKISystemData — Type

DeltaKISystemData{FT<:AbstractFloat}

Represent the internuclear couplings between every pair of molecules, assumed to be computed via a Morse potential, with the possibility of Bloch periodicity, and represent the diagonal elements of that internuclear spring constant matrix for the full system based on the off-diagonal blocks.

See also: DeltaKIMolBlockData{FT}

Morse potential harmonic coupling function

This function can be used independently of AARMBEM-C.jl.

keffMorse — Function

keffMorse(r::FT, k0::FT, a0::FT, Einf::FT) where FT<:AbstractFloat

Compute and return the effective Morse spring constant between two particles at distance r (assumed to be nonnegative), with equilibrium spring constant k0 and separation a0 and with dissociation energy Einf. If $U(r)$ is the potential function, then this returns the quantity $(1/(r - a_{0}))\partial U/\partial r$.

Functions to compute $\Delta K_{\mathrm{I}}$

readDeltaKISystemData — Function

readDeltaKISystemData(mollistfilename::AbstractString,
                      myAllMolData::MolSystem{OneMol_WithPhonons{FT}}) where FT<:AbstractFloat

Read a configuration file mollistfilename (a string) to extract data detailing which pairs of molecules have nontrivial internuclear couplings between them, handling duplicate entries appropriately.

assembleDeltaKI! — Function

assembleDeltaKI!(DeltaKI::AbstractArray{Complex{FT}, 2},
                 myDeltaKIsystemdata::DeltaKISystemData{FT},
                 myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                 myPeriodicData::PeriodicData{FT},
                 blochk::AbstractArray{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the internuclear spring constant matrix of the collection of molecules (which, at any wavevector blochk, will be a Hermitian positive-definite matrix) for the whole system, including the diagonal contributions, and stamp this into DeltaKI.